+++ app/javascript/active_admin/base.js
@@ -17,3 +17,12 @@ import "./initializers/has-many"
 import "./initializers/per-page"
 import "./initializers/table-checkbox-toggler"
 import "./initializers/tabs"
+
+import ModalDialog from "./lib/modal-dialog";
+
+function modal_dialog(message, inputs, callback) {
+  console.warn("ActiveAdmin.modal_dialog is deprecated in favor of ActiveAdmin.ModalDialog, please update usage.");
+  return ModalDialog(message, inputs, callback);
+}
+
+export { ModalDialog, modal_dialog };