+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug Fixes
 
+* Export ModalDialog component to re-enable client side usage. [#5956] by [@sgara]
 * Use default ActionView options instead of default Formtastic options for DateRangeInput [#5957] by [@mirelon]
 
 ## 2.5.0 [☰](https://github.com/activeadmin/activeadmin/compare/v2.4.0..v2.5.0)
@@ -532,6 +533,7 @@ Please check [0-6-stable] for previous changes.
 [#5931]: https://github.com/activeadmin/activeadmin/pull/5931
 [#5938]: https://github.com/activeadmin/activeadmin/pull/5938
 [#5929]: https://github.com/activeadmin/activeadmin/pull/5929
+[#5956]: https://github.com/activeadmin/activeadmin/pull/5956
 [#5957]: https://github.com/activeadmin/activeadmin/pull/5957
 
 [@5t111111]: https://github.com/5t111111